Add Spark 4.x / Scala 2.13 support (#22)
Cross-build the dialect against two Spark / Scala lines, selected with
`-PsparkProfile` (default `spark3`, unchanged behaviour):

  spark3 -> Scala 2.12 / Spark 3.5.x / Java 8-11 -> spark-dialect-extension_2.12
  spark4 -> Scala 2.13 / Spark 4.x   / Java 17   -> spark-dialect-extension_2.13

Build layout (per maintainer guidance in #22):
- Split the single module into Gradle sub-projects: version-agnostic `common`
  (runtime registry + shared test fixtures), and one implementation module per
  Spark line (`spark35`, `spark41`). Only the profile-compatible module is
  included, and each publishes a single fat jar bundling `common` + its dialect.
  The registry selects the implementation at runtime from the Spark version, so
  one Scala 2.13 artifact serves every Spark 4.x minor release.
- Tests are split per Spark version (spark35/spark41) instead of using in-test
  conditionals; shared infrastructure lives in `common` test fixtures.

Spark 4 specifics:
- Override `JdbcDialect.isObjectNotFoundException` to detect ClickHouse error
  code 60 (UNKNOWN_TABLE) via the JDBC `SQLException` vendor code (locale
  independent), so writes to a not-yet-existing table work.
- Add the standard `--add-opens` JVM args to the test task on Java 9+.
- Version-specific array-read expectations: several `Array(Int16/Int32/Int64/
  Float32/Float64)` now read successfully with driver 0.7.x; the rest surface as
  `COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH`.
- Scalafix is limited to the spark3 profile (no semanticdb-scalac for the Scala
  2.13 patch Spark 4 pulls in); Scalafmt runs on both.

Extract `ClickhouseDialectRegistry.dialectClassName` so the version dispatch is
unit-testable without an active Spark session (Spark 4's `getActiveSession`
filters out non-usable/mocked sessions).

CI test and release workflows run both profiles with the matching Java version.
