+++ .github/workflows/release.yml
@@ -11,23 +11,35 @@ jobs:
     permissions:
       contents: write # to create Github release
 
+    strategy:
+      # Publish both artifacts even if one fails; each job appends its jars to the release.
+      fail-fast: false
+      max-parallel: 1
+      matrix:
+        spark-profile: ['spark3', 'spark4']
+        include:
+          - spark-profile: 'spark3'  # Scala 2.12 -> spark-dialect-extension_2.12
+            java: '11'
+          - spark-profile: 'spark4'  # Scala 2.13 -> spark-dialect-extension_2.13
+            java: '17'
+
     steps:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
 
-      - name: Setup JDK 11
+      - name: Setup JDK ${{ matrix.java }}
         uses: actions/setup-java@v4
         with:
-          java-version: '11'
+          java-version: '${{ matrix.java }}'
           distribution: 'temurin'
 
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v3
 
-      - run: ./gradlew jar sourcesJar scaladocJar
+      - run: ./gradlew jar sourcesJar scaladocJar -PsparkProfile=${{ matrix.spark-profile }}
 
-      - run: ./gradlew publishAndReleaseToMavenCentral --no-configuration-cache
+      - run: ./gradlew publishAndReleaseToMavenCentral --no-configuration-cache -PsparkProfile=${{ matrix.spark-profile }}
         env:
           # https://vanniktech.github.io/gradle-maven-publish-plugin/central/#in-memory-gpg-key
           ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.SONATYPE_USERNAME }}
