+++ .scalafmt.conf
@@ -16,5 +16,9 @@ fileOverride {
   "glob:**/src/**/scala-2.13/**.scala" {
     runner.dialect = scala213
   }
+  # Scala 2.13 sources of the Spark 4 implementation module
+  "glob:**/spark41/**.scala" {
+    runner.dialect = scala213
+  }
 }
 version = 3.7.5
