+++ build.gradle
@@ -1,136 +1,94 @@
-import com.vanniktech.maven.publish.SonatypeHost
-
 plugins {
-    id "scala"
-    id "java-library"
-    id "jacoco"
-    id "com.vanniktech.maven.publish" version "0.29.0"
-    id "net.nemerosa.versioning" version "3.1.0"
-    // Scala support
-    id "com.github.maiflai.scalatest" version "0.32"
-    id "cz.augi.gradle.scalafmt" version "1.21.3"
-    id "io.github.cosmicsilence.scalafix" version "0.2.2"
+    // Third-party plugins applied to the sub-projects (see subprojects {} below);
+    // declared here so the plugin artifacts are on the classpath for the whole build.
+    // The core `scala` plugin is always on the classpath and is applied directly.
+    id "com.github.maiflai.scalatest" version "0.32" apply false
+    id "cz.augi.gradle.scalafmt" version "1.21.3" apply false
+    id "io.github.cosmicsilence.scalafix" version "0.2.2" apply false
+    id "com.vanniktech.maven.publish" version "0.29.0" apply false
+    id "net.nemerosa.versioning" version "3.1.0" apply false
 }
 
-group = "io.github.mtsongithub.doetl"
-// TODO: crossbuild for Scala 2.13
-archivesBaseName = "spark-dialect-extension_2.12"
-version = "0.0.6"
+// ---------------------------------------------------------------------------
+// Build profile resolution
+// ---------------------------------------------------------------------------
+def sparkProfile = providers.gradleProperty("sparkProfile").getOrElse("spark3")
 
-repositories {
-    mavenCentral()
+def profiles = [
+    spark3: [
+        scalaBinaryVersion: "2.12",
+        scalaVersion      : "2.12.19",
+        sparkVersion      : "3.5.1",
+        javaVersion       : 8,
+        implProject       : ":spark35",
+    ],
+    spark4: [
+        scalaBinaryVersion: "2.13",
+        scalaVersion      : "2.13.17",
+        sparkVersion      : "4.1.2",
+        javaVersion       : 17,
+        implProject       : ":spark41",
+    ],
+]
+
+if (!profiles.containsKey(sparkProfile)) {
+    throw new GradleException(
+        "Unknown sparkProfile '${sparkProfile}'. Supported values: 'spark3', 'spark4'.")
 }
+def profile = profiles[sparkProfile]
 
-dependencies {
-    // Scala support
-    implementation "org.scala-lang:scala-library:2.12.19"
-    // core Spark dependencies
-    // TODO: crossbuild for Spark 3.2-3.5
-    implementation "org.apache.spark:spark-core_2.12:3.5.1"
-    implementation "org.apache.spark:spark-sql_2.12:3.5.1"
-    // latest compatible with Java 8
-    implementation "ch.qos.logback:logback-classic:1.3.14"
+allprojects {
+    group = "io.github.mtsongithub.doetl"
+    version = "0.0.6"
 
-    // test dependencies
-    testImplementation "org.scalatest:scalatest_2.12:3.2.19"
-    testImplementation "com.vladsch.flexmark:flexmark-all:0.64.8"
-    testImplementation "org.mockito:mockito-scala_2.12:1.17.37"
-    // latest compatible with Java 8
-    testImplementation "io.github.cdimascio:dotenv-java:3.1.0"
+    ext {
+        sparkProfileName        = sparkProfile
+        scalaBinaryVersion      = profile.scalaBinaryVersion
+        scalaFullVersion        = profile.scalaVersion
+        sparkVersion            = profile.sparkVersion
+        profileJavaVersion      = profile.javaVersion
+        implProjectPath         = profile.implProject
+        // published artifact name, e.g. spark-dialect-extension_2.12
+        publishedArtifactId     = "spark-dialect-extension_${profile.scalaBinaryVersion}"
+        clickhouseJdbcVersion   = providers.gradleProperty("clickhouse.jdbc.version").getOrElse("0.9.8")
+        isReleaseVersion        = !version.toString().endsWith("SNAPSHOT")
+    }
 
-    // Clickhouse dependency
-    testImplementation ("com.clickhouse:clickhouse-jdbc:${project.properties['clickhouse.jdbc.version']}${project.properties['clickhouse.jdbc.version']?.startsWith('0.9') ? ':all' : ''}") {
-        exclude group: "org.antlr", module: "antlr4-runtime"
-        exclude group: "at.yawk.lz4", module: "lz4-java"
+    repositories {
+        mavenCentral()
     }
 }
 
-ext {
-    isReleaseVersion = !version.endsWith("SNAPSHOT")
-}
+// ---------------------------------------------------------------------------
+// Shared configuration for the Scala modules (:common and the active impl)
+// ---------------------------------------------------------------------------
+subprojects {
+    apply plugin: "scala"
+    apply plugin: "cz.augi.gradle.scalafmt"
 
-test {
-    finalizedBy jacocoTestReport
-    systemProperty("CLICKHOUSE_JDBC_VERSION", project.properties['clickhouse.jdbc.version'])
-}
-
-jacocoTestReport {
-    dependsOn test
-
-    reports {
-        xml.required = true
+    // Scalafix relies on semanticdb-scalac, which is not published for the Scala
+    // 2.13 patch that Spark 4 pulls in. Keep Scalafix on the Scala 2.12 profile
+    // only; Scalafmt runs on both.
+    if (rootProject.ext.scalaBinaryVersion == "2.12") {
+        apply plugin: "io.github.cosmicsilence.scalafix"
     }
-}
 
-tasks.withType(ScalaCompile) {
-    scalaCompileOptions.additionalParameters = ["-Ywarn-unused"]
-}
+    java {
+        targetCompatibility = JavaVersion.toVersion(rootProject.ext.profileJavaVersion)
+    }
 
-getTasksByName("test", true).each { task ->
-    task.dependsOn(jar)
+    dependencies {
+        implementation "org.scala-lang:scala-library:${rootProject.ext.scalaFullVersion}"
+        implementation "org.apache.spark:spark-sql_${rootProject.ext.scalaBinaryVersion}:${rootProject.ext.sparkVersion}"
+    }
+
+    tasks.withType(ScalaCompile).configureEach {
+        scalaCompileOptions.additionalParameters = ["-Ywarn-unused"]
+    }
 }
 
 task printVersion {
     doLast {
         println project.version
-    }
-}
-
-task scaladocJar(type: Jar, dependsOn: scaladoc) {
-    archiveClassifier = "scaladoc"
-    from scaladoc.destinationDir
-}
-
-java {
-    targetCompatibility = JavaVersion.VERSION_1_8
-    withSourcesJar()
-}
-
-jar {  
-    manifest {
-        attributes(
-            "Implementation-Title": project.name,
-            "Implementation-Version": project.version,
-            "Implementation-Vendor": "MTS PJSC",
-            "Created-By": "Gradle ${gradle.gradleVersion}",
-            "Built-By": System.properties["user.name"],
-            "Build-Timestamp": versioning.info.time,
-            "Build-Revision": versioning.info.commit,
-            "Build-Jdk": "${System.properties["java.version"]} (${System.properties["java.vendor"]} ${System.properties["java.vm.version"]})",
-            "Build-OS": "${System.properties["os.name"]} ${System.properties["os.arch"]} ${System.properties["os.version"]}",
-        )
-    }
-}
-
-mavenPublishing {
-    publishToMavenCentral(SonatypeHost.CENTRAL_PORTAL)
-    signAllPublications()
-
-    coordinates(project.group, "spark-dialect-extension_2.12", project.version)
-
-    pom {
-        name = "spark-dialect-extension"
-        description = "Spark dialect extension for enhanced type handling"
-        url = "https://mws.ru/data/mws-data-bridge/"
-        inceptionYear = "2024"
-        licenses {
-            license {
-                name = "Apache-2.0"
-                url = "http://www.apache.org/licenses/LICENSE-2.0"
-                distribution = "http://www.apache.org/licenses/LICENSE-2.0.txt"
-            }
-        }
-        developers {
-            developer {
-                id = "doetl"
-                name = "MWS Data Bridge"
-                email = "onetools@mts.ru"
-            }
-        }
-        scm {
-            connection = "scm:git:git://github.com/MTSWebServices/spark-dialect-extension.git"
-            developerConnection = "scm:git:ssh://github.com/MTSWebServices/spark-dialect-extension.git"
-            url = "https://github.com/MTSWebServices/spark-dialect-extension"
-        }
     }
 }
