+++ gradle.properties
@@ -1,1 +1,7 @@
+# Build profile selecting the Spark / Scala axis:
+#   spark3 -> Scala 2.12 / Spark 3.5.x / Java 8-11 (default, current behaviour)
+#   spark4 -> Scala 2.13 / Spark 4.x   / Java 17
+# Override on the command line with -PsparkProfile=spark4
+sparkProfile=spark3
+
 clickhouse.jdbc.version=0.9.8
