+++ spark41/src/test/scala/io/github/mtsongithub/doetl/sparkdialectextensions/clickhouse/ClickhouseDialectTest.scala
@@ -1,0 +1,1092 @@
+package io.github.mtsongithub.doetl.sparkdialectextensions.clickhouse
+
+import io.github.mtsongithub.doetl.sparkdialectextensions.SharedSparkSession
+import org.apache.spark.sql.Row
+import org.apache.spark.sql.jdbc.JdbcDialects
+import org.apache.spark.sql.types._
+import org.scalatest.funsuite.AnyFunSuite
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.prop.TableDrivenPropertyChecks
+
+class ClickhouseDialectTest
+    extends AnyFunSuite
+    with Matchers
+    with SharedSparkSession
+    with TableDrivenPropertyChecks
+    with ClickhouseFixture {
+
+  /**
+   * this function is used to normalize schemas for comparison purposes to handle cases such as:
+   * 'StructField("shortColumn", ShortType, true, {"scale":0})' is equal to
+   * 'StructField("shortColumn", ShortType, true, {})'
+   */
+  def stripMetadata(schema: StructType): StructType = {
+    StructType(schema.fields.map(f => f.copy(metadata = Metadata.empty)))
+  }
+
+  test("custom clickhouse dialect usage") {
+    val dialect = JdbcDialects.get("jdbc:clickhouse")
+    assert(dialect.getClass.getName.contains("ClickhouseDialectExtension"))
+  }
+
+  test("throws exception for unsupported spark version") {
+    val exception = intercept[UnsupportedOperationException] {
+      ClickhouseDialectRegistry.dialectClassName("3.4.0")
+    }
+    // check the exception message
+    exception.getMessage should include("Unsupported Spark version: 3.4.0")
+  }
+
+  test("test clickHouse table read with spark") {
+
+    val schema =
+      """
+        |booleanColumn Boolean,
+        |byteColumn Int8,
+        |shortColumn Int16,
+        |timestampColumn Datetime64(6),
+    """.stripMargin
+    setupTable(schema)
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    val expectedSchema = StructType(
+      Seq(
+        StructField("booleanColumn", BooleanType, nullable = true),
+        StructField("byteColumn", ByteType, nullable = true),
+        StructField("shortColumn", ShortType, nullable = true),
+        StructField("timestampColumn", TimestampType, nullable = true)))
+
+    assert(df.schema.treeString === expectedSchema.treeString)
+  }
+
+  test("test spark dataframe write to existing clickhouse table") {
+
+    val schema =
+      """
+        |booleanColumn Boolean,
+        |byteColumn Int8,
+        |shortColumn Int16,
+        |timestampColumn Datetime64(6),
+    """.stripMargin
+    setupTable(schema)
+
+    val generator = new ClickhouseDataframeGenerator(spark)
+    val df = generator.createDataFrame()
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .mode("append")
+      .save()
+
+    val loadedDf = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(stripMetadata(loadedDf.schema) == stripMetadata(df.schema))
+  }
+
+  test("test spark dataframe write to new clickhouse table") {
+
+    val generator = new ClickhouseDataframeGenerator(spark)
+    val df = generator.createDataFrame()
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val loadedDf = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(stripMetadata(loadedDf.schema) == stripMetadata(df.schema))
+  }
+
+  test("test spark dataframe write to existing clickhouse table with truncate option") {
+
+    val generator = new ClickhouseDataframeGenerator(spark)
+    val df = generator.createDataFrame()
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("createTableOptions", "ENGINE = Log")
+      .mode("errorIfExists")
+      .save()
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("truncate", "true")
+      .mode("overwrite")
+      .save()
+
+    val loadedDf = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(getTableEngine(tableName) == "Log")
+    assert(stripMetadata(loadedDf.schema) == stripMetadata(df.schema))
+    assert(loadedDf.count() == df.count())
+  }
+
+  test("read ClickHouse Int8 as Spark ByteType") {
+    setupTable("byteColumn Int8")
+    insertTestData(Seq("(-128)", "(127)")) // min and max values for a signed byte
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == ByteType)
+
+    val data = df.collect().map(_.getByte(0)).sorted
+    assert(data sameElements Array(-128, 127))
+  }
+
+  test("read ClickHouse Int16 as Spark ShortType") {
+    setupTable("shortColumn Int16")
+    insertTestData(Seq("(-32768)", "(32767)")) // min and max values for a signed short
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == ShortType)
+
+    val data = df.collect().map(_.getShort(0)).sorted
+    assert(data sameElements Array(-32768, 32767))
+  }
+
+  test("read ClickHouse Int32 as Spark IntegerType") {
+    setupTable("integerColumn Int32")
+    insertTestData(
+      Seq("(-2147483648)", "(2147483647)")
+    ) // min and max values for a signed integer
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == IntegerType)
+
+    val data = df.collect().map(_.getInt(0)).sorted
+    assert(data sameElements Array(-2147483648, 2147483647))
+  }
+
+  test("read ClickHouse Int64 as Spark LongType") {
+    setupTable("longColumn Int64")
+    insertTestData(
+      Seq("(-9223372036854775808)", "(9223372036854775807)")
+    ) // min and max values for a signed long
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == LongType)
+
+    val data = df.collect().map(_.getLong(0)).sorted
+    assert(data sameElements Array(-9223372036854775808L, 9223372036854775807L))
+  }
+
+  test("write Spark ShortType as ClickHouse Int16") {
+    val schema = StructType(Seq(StructField("shortColumn", ShortType, nullable = false)))
+    val data = Seq(Row(Short.MinValue), Row(Short.MaxValue))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val statement = connection.createStatement()
+    val rs = statement.executeQuery(s"DESCRIBE TABLE $tableName")
+    while (rs.next()) {
+      val columnName = rs.getString("name")
+      val dataType = rs.getString("type")
+      columnName match {
+        case "shortColumn" => assert(dataType == "Int16")
+      }
+    }
+    statement.close()
+  }
+
+  test("read ClickHouse UInt8 as Spark ShortType") {
+    setupTable("uByteColumn UInt8")
+    insertTestData(Seq("(0)", "(255)")) // min and max values for unsigned byte
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == ShortType)
+
+    val data = df.collect().map(_.getShort(0)).sorted
+    assert(data sameElements Array(0, 255))
+  }
+
+  test("read ClickHouse UInt16 as Spark IntegerType") {
+    setupTable("uShortColumn UInt16")
+    insertTestData(Seq("(0)", "(65535)")) // min and max values for unsigned short
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == IntegerType)
+
+    val data = df.collect().map(_.getInt(0)).sorted
+    assert(data sameElements Array(0, 65535))
+  }
+
+  test("read ClickHouse UInt32 as Spark LongType") {
+    setupTable("uIntColumn UInt32")
+    insertTestData(Seq("(0)", "(4294967295)")) // min and max values for unsigned int
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == LongType)
+
+    val data = df.collect().map(_.getLong(0)).sorted
+    assert(data sameElements Array(0L, 4294967295L))
+  }
+
+  test("read Clickhouse UInt64 as Spark DecimalType") {
+    setupTable("UInt64Column UInt64")
+    insertTestData(Seq("(0)", "(18446744073709551615)")) // min and max values
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DecimalType(38, 0))
+
+    val data = df.collect().map(_.getDecimal(0)).sorted
+    assert(
+      data sameElements Array(
+        new java.math.BigDecimal(0),
+        new java.math.BigDecimal("18446744073709551615")))
+  }
+
+  test("read ClickHouse Float32 as Spark FloatType") {
+    setupTable("floatColumn Float32")
+    insertTestData(Seq("(-1.23)", "(4.56)"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == FloatType)
+    val data = df.collect().map(_.getFloat(0)).sorted
+    assert(data sameElements Array(-1.23f, 4.56f))
+  }
+
+  test("read ClickHouse Float64 as Spark DoubleType") {
+    setupTable("doubleColumn Float64")
+    insertTestData(
+      Seq("(-1.7976931348623157E308)", "(1.7976931348623157E308)")
+    ) // min and max values for double
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DoubleType)
+
+    val data = df.collect().map(_.getDouble(0)).sorted
+    assert(data sameElements Array(-1.7976931348623157e308, 1.7976931348623157e308))
+  }
+
+  test("read ClickHouse Decimal(9,2) as Spark DecimalType") {
+    setupTable("decimalColumn Decimal(9,2)")
+    insertTestData(Seq("(12345.67)", "(89012.34)"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DecimalType(9, 2))
+
+    val data = df.collect().map(_.getDecimal(0)).sorted
+    assert(
+      data sameElements Array(
+        new java.math.BigDecimal("12345.67"),
+        new java.math.BigDecimal("89012.34")))
+  }
+
+  test("read ClickHouse Decimal32 as Spark DecimalType(9, scale)") {
+    setupTable("decimalColumn Decimal32(2)")
+    insertTestData(Seq("(12345.67)", "(89012.34)"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DecimalType(9, 2))
+
+    val data = df.collect().map(_.getDecimal(0)).sorted
+    assert(
+      data sameElements Array(
+        new java.math.BigDecimal("12345.67"),
+        new java.math.BigDecimal("89012.34")))
+  }
+
+  test("read ClickHouse Decimal64 as Spark DecimalType(18, scale)") {
+    setupTable("decimalColumn Decimal64(2)")
+    insertTestData(Seq("(123456789.12)", "(987654321.34)"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DecimalType(18, 2))
+
+    val data = df.collect().map(_.getDecimal(0)).sorted
+    assert(
+      data sameElements Array(
+        new java.math.BigDecimal("123456789.12"),
+        new java.math.BigDecimal("987654321.34")))
+  }
+
+  test("read ClickHouse Decimal128 as Spark DecimalType(38, scale)") {
+    setupTable("decimalColumn Decimal128(2)")
+    insertTestData(
+      Seq("(123456789012345678901234567890.12)", "(987654321098765432109876543210.34)"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DecimalType(38, 2))
+
+    val data = df.collect().map(_.getDecimal(0)).sorted
+    assert(
+      data sameElements Array(
+        new java.math.BigDecimal("123456789012345678901234567890.12"),
+        new java.math.BigDecimal("987654321098765432109876543210.34")))
+  }
+
+  test("read ClickHouse Date as Spark DateType") {
+    setupTable("dateColumn Date")
+    insertTestData(Seq("('2023-01-01')", "('2023-12-31')"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == DateType)
+
+    val data = df.collect().map(_.getDate(0))
+    assert(
+      data sameElements Array(
+        java.sql.Date.valueOf("2023-01-01"),
+        java.sql.Date.valueOf("2023-12-31")))
+  }
+
+  test("read ClickHouse IPv4 as Spark StringType") {
+    setupTable("ipv4Column IPv4")
+    insertTestData(Seq("('0.0.0.0')", "('255.255.255.255')"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == StringType)
+
+    val data = df.collect().map(_.getString(0)).sorted
+    assert(data sameElements Array("0.0.0.0", "255.255.255.255"))
+  }
+
+  test("read ClickHouse IPv6 as Spark StringType") {
+    setupTable("ipv6Column IPv6")
+    insertTestData(Seq("('2001:db8::42')", "('dbd6:a34d:7835:c75:304f:6c73:edac:957b')"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == StringType)
+
+    val data = df.collect().map(_.getString(0)).sorted
+    assert(
+      data sameElements Array("2001:db8:0:0:0:0:0:42", "dbd6:a34d:7835:c75:304f:6c73:edac:957b"))
+  }
+
+  test("read ClickHouse UUID as Spark StringType") {
+    setupTable("uuidColumn UUID")
+    insertTestData(
+      Seq("('483971cf-aad7-4c84-abf1-4a94c9d72f99')", "('8f3c71a9-b4d2-40e1-95c8-63a2b74051f9')"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == StringType)
+
+    val data = df.collect().map(_.getString(0)).sorted
+    assert(
+      data sameElements Array(
+        "483971cf-aad7-4c84-abf1-4a94c9d72f99",
+        "8f3c71a9-b4d2-40e1-95c8-63a2b74051f9"))
+  }
+
+  test("read ClickHouse DateTime as Spark TimestampType") {
+    setupTable("datetimeColumn DateTime")
+    insertTestData(Seq("('2023-01-01 12:34:56')", "('2023-12-31 23:59:59')"))
+
+    val df = spark.read
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .load()
+
+    assert(df.schema.fields.head.dataType == TimestampType)
+
+    val data = df.collect().map(_.getTimestamp(0))
+    assert(
+      data sameElements Array(
+        java.sql.Timestamp.valueOf("2023-01-01 12:34:56"),
+        java.sql.Timestamp.valueOf("2023-12-31 23:59:59")))
+  }
+
+  test("write Spark TimestampType as ClickHouse Datetime64(6)") {
+    val schema = StructType(Seq(StructField("timestampColumn", TimestampType, nullable = false)))
+    val currentTime = new java.sql.Timestamp(System.currentTimeMillis())
+    val data = Seq(Row(currentTime))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val statement = connection.createStatement()
+    val rs = statement.executeQuery(s"DESCRIBE TABLE $tableName")
+    while (rs.next()) {
+      val columnName = rs.getString("name")
+      val dataType = rs.getString("type")
+      columnName match {
+        case "timestampColumn" => assert(dataType == "DateTime64(6)")
+      }
+    }
+    statement.close()
+  }
+
+  test("write Spark BooleanType as ClickHouse Bool") {
+    val schema = StructType(Seq(StructField("booleanColumn", BooleanType, nullable = false)))
+    val data = Seq(Row(true), Row(false))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val statement = connection.createStatement()
+    val rs = statement.executeQuery(s"DESCRIBE TABLE $tableName")
+    while (rs.next()) {
+      val columnName = rs.getString("name")
+      val dataType = rs.getString("type")
+      columnName match {
+        case "booleanColumn" => assert(dataType == "Bool")
+      }
+    }
+    statement.close()
+  }
+
+  test("write Spark Nullable(ShortType) as ClickHouse Nullable(Int16)") {
+    val schema = StructType(Seq(StructField("nullableShortColumn", ShortType, nullable = true)))
+    val data = Seq(Row(null), Row(Short.MinValue), Row(Short.MaxValue))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val actualColumnType = getColumnType("nullableShortColumn")
+    assert(actualColumnType == "Nullable(Int16)")
+  }
+
+  test("write Spark Nullable(TimestampType) as ClickHouse Nullable(Datetime64(6))") {
+    val schema =
+      StructType(Seq(StructField("nullableTimestampColumn", TimestampType, nullable = true)))
+    val currentTime = new java.sql.Timestamp(System.currentTimeMillis())
+    val data = Seq(Row(null), Row(currentTime))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val actualColumnType = getColumnType("nullableTimestampColumn")
+    assert(actualColumnType == "Nullable(DateTime64(6))")
+  }
+
+  test("write Spark Nullable(BooleanType) as ClickHouse Nullable(Bool)") {
+    val schema =
+      StructType(Seq(StructField("nullableBooleanColumn", BooleanType, nullable = true)))
+    val data = Seq(Row(null), Row(true), Row(false))
+    val df = spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+
+    df.write
+      .format("jdbc")
+      .option("url", jdbcUrl)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("dbtable", tableName)
+      .option("user", jdbcUser)
+      .option("password", jdbcPassword)
+      .option("createTableOptions", "ENGINE = TinyLog")
+      .mode("errorIfExists")
+      .save()
+
+    val actualColumnType = getColumnType("nullableBooleanColumn")
+    assert(actualColumnType == "Nullable(Bool)")
+  }
+
+  val testReadArrayCasesV0_9_X = Table(
+    ("columnDefinition", "insertedData", "expectedType"),
+    (
+      "charArrayColumn Array(String)",
+      "(['a', 'b', 'c', 'd', 'e'])",
+      ArrayType(StringType, containsNull = false)),
+    (
+      "ipv4ArrayColumn Array(IPv4)",
+      "(['0.0.0.0', '192.168.1.1', '255.255.255.255'])",
+      ArrayType(StringType, containsNull = false)),
+    (
+      "ipv6ArrayColumn Array(IPv6)",
+      "(['2001:db8::42', 'dbd6:a34d:7835:c75:304f:6c73:edac:957b'])",
+      ArrayType(StringType, containsNull = false)),
+    (
+      "uuidArrayColumn Array(UUID)",
+      "(['483971cf-aad7-4c84-abf1-4a94c9d72f99', '8f3c71a9-b4d2-40e1-95c8-63a2b74051f9'])",
+      ArrayType(StringType, containsNull = false)),
+    (
+      "byteArrayColumn Array(Int8)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ByteType, containsNull = false)),
+    (
+      "shortArrayColumn Array(Int16)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ShortType, containsNull = false)),
+    (
+      "intArrayColumn Array(Int32)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(IntegerType, containsNull = false)),
+    (
+      "longArrayColumn Array(Int64)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(LongType, containsNull = false)),
+    (
+      "doubleArrayColumn Array(Float64)",
+      "([1.1, 2.2, 3.3, 4.4, 5.5])",
+      ArrayType(DoubleType, containsNull = false)),
+    (
+      "shortArrayColumn Array(UInt8)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ShortType, containsNull = false)),
+    (
+      "intArrayColumn Array(UInt16)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(IntegerType, containsNull = false)),
+    (
+      "longArrayColumn Array(UInt32)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(LongType, containsNull = false)),
+    // since 0.9.5 https://github.com/ClickHouse/clickhouse-java/pull/2627
+    (
+      "floatColumn Array(Float32)",
+      "([(1.1), (2.2), (3.3), (4.4), (5.5)])",
+      ArrayType(FloatType, containsNull = false)),
+    (
+      "decimalArrayColumn Array(Decimal(9,2))",
+      "([1.23, 2.34, 3.45, 4.56, 5.67])",
+      ArrayType(DecimalType(9, 2), containsNull = false)),
+    // since 0.8.6 https://github.com/ClickHouse/clickhouse-java/issues/1409
+    (
+      "dateArrayColumn Array(Date)",
+      "(['2024-01-01', '2024-01-02', '2024-01-03'])",
+      ArrayType(DateType, containsNull = false)),
+    (
+      "datetimeArrayColumn Array(DateTime64(6))",
+      "(['2024-01-01T00:00:00.000000', '2024-01-02T11:11:11.111111', '2024-01-03.2222222'])",
+      ArrayType(TimestampType, containsNull = false)))
+
+  val testReadArrayCasesV0_7_X = Table(
+    ("columnDefinition", "insertedData", "expectedType"),
+    (
+      "charArrayColumn Array(String)",
+      "(['a', 'b', 'c', 'd', 'e'])",
+      ArrayType(StringType, containsNull = false)),
+    (
+      "decimalArrayColumn Array(Decimal(9,2))",
+      "([1.23, 2.34, 3.45, 4.56, 5.67])",
+      ArrayType(DecimalType(9, 2), containsNull = false)),
+    // On Spark 4 these signed numeric arrays read successfully even with driver 0.7.x,
+    // unlike on Spark 3.5 where they raised a ClassCastException.
+    (
+      "shortArrayColumn Array(Int16)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ShortType, containsNull = false)),
+    (
+      "intArrayColumn Array(Int32)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(IntegerType, containsNull = false)),
+    (
+      "longArrayColumn Array(Int64)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(LongType, containsNull = false)),
+    (
+      "floatArrayColumn Array(Float32)",
+      "([1.1, 2.2, 3.3, 4.4, 5.5])",
+      ArrayType(FloatType, containsNull = false)),
+    (
+      "doubleArrayColumn Array(Float64)",
+      "([1.1, 2.2, 3.3, 4.4, 5.5])",
+      ArrayType(DoubleType, containsNull = false)))
+
+  forAll(
+    if (driverVersion.startsWith("0.9")) testReadArrayCasesV0_9_X else testReadArrayCasesV0_7_X) {
+    (columnDefinition: String, insertedData: String, expectedType: DataType) =>
+      test(s"read ClickHouse Array for ${columnDefinition} column") {
+        setupTable(columnDefinition)
+        insertTestData(Seq(insertedData))
+
+        val df = spark.read
+          .format("jdbc")
+          .option("url", jdbcUrl)
+          .option("user", jdbcUser)
+          .option("password", jdbcPassword)
+          .option("dbtable", tableName)
+          .load()
+
+        assert(df.schema.fields.head.dataType === expectedType)
+
+        // In Scala 2.13 `Row.getAs[Seq[_]]` returns a mutable ArraySeq, which is not an
+        // immutable Seq; use the collection-root Seq to stay version-neutral.
+        df.collect().map(_.getAs[scala.collection.Seq[Any]](df.schema.fields.head.name)).head
+        assert(df.count() == 1)
+      }
+  }
+
+  // On Spark 4 these still fail, but as a COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH error instead of
+  // the driver-level ClassCastException seen on Spark 3.5. The signed numeric types that used to
+  // be here (Int16/Int32/Int64/Float32/Float64) now read successfully, see testReadArrayCasesV0_7_X.
+  val testReadArrayUnsupportedCasesV0_7_X = Table(
+    ("columnDefinition", "insertedData", "expectedType", "errorMessage"),
+    // https://github.com/ClickHouse/clickhouse-java/issues/1754
+    (
+      "byteArrayColumn Array(Int8)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ByteType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    (
+      "shortArrayColumn Array(UInt8)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(ShortType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    (
+      "intArrayColumn Array(UInt16)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(IntegerType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    (
+      "longArrayColumn Array(UInt32)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(LongType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    (
+      "UInt64Column Array(UInt64)",
+      "([1, 2, 3, 4, 5])",
+      ArrayType(DecimalType(38, 0), containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    // https://github.com/ClickHouse/clickhouse-java/issues/1409
+    (
+      "dateArrayColumn Array(Date)",
+      "(['2024-01-01', '2024-01-02', '2024-01-03'])",
+      ArrayType(DateType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"),
+    (
+      "datetimeArrayColumn Array(DateTime64(6))",
+      "(['2024-01-01T00:00:00.000000', '2024-01-02T11:11:11.111111', '2024-01-03.2222222'])",
+      ArrayType(TimestampType, containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"))
+
+  val testReadArrayUnsupportedCasesV0_9_X = Table(
+    ("columnDefinition", "insertedData", "expectedType", "errorMessage"),
+    (
+      "UInt64Column Array(UInt64)",
+      "([(1), (2), (3), (4), (5)])",
+      ArrayType(DecimalType(38, 0), containsNull = false),
+      "COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH"))
+
+  forAll(
+    if (driverVersion.startsWith("0.9")) testReadArrayUnsupportedCasesV0_9_X
+    else testReadArrayUnsupportedCasesV0_7_X) {
+    (
+        columnDefinition: String,
+        insertedData: String,
+        expectedType: DataType,
+        errorMessage: String) =>
+      test(s"cannot read ClickHouse Array for ${columnDefinition} column") {
+        setupTable(columnDefinition)
+        insertTestData(Seq(insertedData))
+
+        // schema is detected properly
+        val df = spark.read
+          .format("jdbc")
+          .option("url", jdbcUrl)
+          .option("user", jdbcUser)
+          .option("password", jdbcPassword)
+          .option("dbtable", tableName)
+          .load()
+
+        assert(df.schema.fields.head.dataType === expectedType)
+
+        // but read is failing (on Spark 4 the driver-level ClassCastException surfaces as a
+        // SparkSQLException with error class COLUMN_ARRAY_ELEMENT_TYPE_MISMATCH)
+        val exception = intercept[Exception] {
+          df.collect()
+        }
+        // check the exception message
+        exception.getMessage should include(errorMessage)
+      }
+  }
+
+  val testWriteArrayCases = Table(
+    ("columnName", "insertedData", "expectedType", "expectedClickhouseType", "nullable"),
+    (
+      "nullableCharArrayColumn",
+      Seq(Row(Array(null, "a", "b", "c", "d", "e"))),
+      ArrayType(StringType, containsNull = true),
+      "Array(Nullable(String))",
+      true),
+    (
+      "nullableByteArrayColumn",
+      Seq(Row(Array(null, 1.toByte, 2.toByte, 3.toByte, 4.toByte, 5.toByte))),
+      ArrayType(ByteType, containsNull = true),
+      "Array(Nullable(Int8))",
+      true),
+    (
+      "nullableShortArrayColumn",
+      Seq(Row(Array(null, 1.toShort, 2.toShort, 3.toShort, 4.toShort, 5.toShort))),
+      ArrayType(ShortType, containsNull = true),
+      "Array(Nullable(Int16))",
+      true),
+    (
+      "nullableIntArrayColumn",
+      Seq(Row(Array(null, 1, 2, 3, 4, 5))),
+      ArrayType(IntegerType, containsNull = true),
+      "Array(Nullable(Int32))",
+      true),
+    (
+      "nullableLongArrayColumn",
+      Seq(Row(Array(null, 1L, 2L, 3L, 4L, 5L))),
+      ArrayType(LongType, containsNull = true),
+      "Array(Nullable(Int64))",
+      true),
+    (
+      "nullableFloatArrayColumn",
+      Seq(Row(Array(null, 1.0f, 2.0f, 3.0f, 4.0f, 5.0f))),
+      ArrayType(FloatType, containsNull = true),
+      "Array(Nullable(Float32))",
+      true),
+    (
+      "nullableDoubleArrayColumn",
+      Seq(Row(Array(null, 1.0d, 2.0d, 3.0d, 4.0d, 5.0d))),
+      ArrayType(DoubleType, containsNull = true),
+      "Array(Nullable(Float64))",
+      true),
+    (
+      "nullableDecimalArrayColumn",
+      Seq(
+        Row(Array(
+          null,
+          new java.math.BigDecimal("1.23"),
+          new java.math.BigDecimal("2.34"),
+          new java.math.BigDecimal("3.45"),
+          new java.math.BigDecimal("4.56"),
+          new java.math.BigDecimal("5.67")))),
+      ArrayType(DecimalType(9, 2), containsNull = true),
+      "Array(Nullable(Decimal(9, 2)))",
+      true),
+    (
+      "nullableDateArrayColumn",
+      Seq(
+        Row(
+          Array(
+            null,
+            java.sql.Date.valueOf("2022-01-01"),
+            java.sql.Date.valueOf("2022-01-02"),
+            java.sql.Date.valueOf("2022-01-03")))),
+      ArrayType(DateType, containsNull = true),
+      "Array(Nullable(Date))",
+      true),
+    (
+      "nullableDatetimeArrayColumn",
+      Seq(
+        Row(Array(
+          null,
+          java.sql.Timestamp.valueOf("2022-01-01 00:00:00.000000"),
+          java.sql.Timestamp.valueOf("2022-01-02 11:11:11.111111"),
+          java.sql.Timestamp.valueOf("2022-01-03 22:22:22.222222")))),
+      ArrayType(TimestampType, containsNull = true),
+      "Array(Nullable(DateTime64(6)))",
+      true),
+    (
+      "charArrayColumn",
+      Seq(Row(Array("a", "b", "c", "d", "e"))),
+      ArrayType(StringType, containsNull = false),
+      "Array(String)",
+      false),
+    (
+      "byteArrayColumn",
+      Seq(Row(Array(1.toByte, 2.toByte, 3.toByte, 4.toByte, 5.toByte))),
+      ArrayType(ByteType, containsNull = false),
+      "Array(Int8)",
+      false),
+    (
+      "shortArrayColumn",
+      Seq(Row(Array(1.toShort, 2.toShort, 3.toShort, 4.toShort, 5.toShort))),
+      ArrayType(ShortType, containsNull = false),
+      "Array(Int16)",
+      false),
+    (
+      "intArrayColumn",
+      Seq(Row(Array(1, 2, 3, 4, 5))),
+      ArrayType(IntegerType, containsNull = false),
+      "Array(Int32)",
+      false),
+    (
+      "longArrayColumn",
+      Seq(Row(Array(1L, 2L, 3L, 4L, 5L))),
+      ArrayType(LongType, containsNull = false),
+      "Array(Int64)",
+      false),
+    (
+      "floatArrayColumn",
+      Seq(Row(Array(1.0f, 2.0f, 3.0f, 4.0f, 5.0f))),
+      ArrayType(FloatType, containsNull = false),
+      "Array(Float32)",
+      false),
+    (
+      "doubleArrayColumn",
+      Seq(Row(Array(1.0d, 2.0d, 3.0d, 4.0d, 5.0d))),
+      ArrayType(DoubleType, containsNull = false),
+      "Array(Float64)",
+      false),
+    (
+      "decimalArrayColumn",
+      Seq(
+        Row(Array(
+          new java.math.BigDecimal("1.23"),
+          new java.math.BigDecimal("2.34"),
+          new java.math.BigDecimal("3.45"),
+          new java.math.BigDecimal("4.56"),
+          new java.math.BigDecimal("5.67")))),
+      ArrayType(DecimalType(9, 2), containsNull = false),
+      "Array(Decimal(9, 2))",
+      false),
+    (
+      "dateArrayColumn",
+      Seq(
+        Row(
+          Array(
+            java.sql.Date.valueOf("2022-01-01"),
+            java.sql.Date.valueOf("2022-01-02"),
+            java.sql.Date.valueOf("2022-01-03")))),
+      ArrayType(DateType, containsNull = false),
+      "Array(Date)",
+      false),
+    (
+      "datetimeArrayColumn",
+      Seq(
+        Row(Array(
+          java.sql.Timestamp.valueOf("2022-01-01 00:00:00.000000"),
+          java.sql.Timestamp.valueOf("2022-01-02 11:11:11.111111"),
+          java.sql.Timestamp.valueOf("2022-01-03 22:22:22.222222")))),
+      ArrayType(TimestampType, containsNull = false),
+      "Array(DateTime64(6))",
+      false))
+
+  forAll(testWriteArrayCases) {
+    (
+        columnName: String,
+        insertedData: Seq[Row],
+        expectedType: DataType,
+        expectedClickhouseType: String,
+        nullable: Boolean) =>
+      test(
+        s"write ClickHouse Array for  ${if (nullable) s" ${columnName} Nullable" else columnName} column") {
+
+        val schema = StructType(Array(StructField(columnName, expectedType, nullable = nullable)))
+        val df = spark.createDataFrame(spark.sparkContext.parallelize(insertedData), schema)
+
+        df.write
+          .format("jdbc")
+          .option("url", jdbcUrl)
+          .option("user", jdbcUser)
+          .option("password", jdbcPassword)
+          .option("dbtable", tableName)
+          .option("createTableOptions", "ENGINE = TinyLog")
+          .mode("errorIfExists")
+          .save()
+
+        val actualColumnType = getColumnType(columnName)
+        assert(actualColumnType == expectedClickhouseType)
+      }
+  }
+}
